Promote inner test classes stubs to first level classes
