+++ src/test/java/de/bechte/junit/stubs/statements/rules/TestWithMethodRuleAndTwoHierarchies.java
@@ -1,0 +1,20 @@
+package de.bechte.junit.stubs.statements.rules;
+
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.MethodRule;
+
+public class TestWithMethodRuleAndTwoHierarchies {
+    @Rule
+    public MethodRule rule;
+
+    public TestWithMethodRuleAndTwoHierarchies(CapturingMethodRuleStub capturingMethodRuleStub) {
+        rule = capturingMethodRuleStub;
+    }
+
+    public class Context {
+        @Test
+        public void aTest() {
+        }
+    }
+}
