+++ src/test/java/de/bechte/junit/stubs/statements/rules/TestWithoutRule.java
@@ -1,0 +1,11 @@
+package de.bechte.junit.stubs.statements.rules;
+
+import org.junit.Test;
+
+public class TestWithoutRule {
+    public class Context {
+        @Test
+        public void aTest() {
+        }
+    }
+}
